fix ошибки по PR
