+++ bot/help/enums.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class DeviceEnum(str, Enum):
+    ANDROID = "android"
+    IOS = "ios"
+    PC = "pc"
+    TV = "tv"
