+++ bot/users/enums.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class ChatType(str, Enum):
+    """Типы чатов в Telegram."""
+
+    PRIVATE = "private"
